requirements.txt
pvporcupine
pvcobra
pyaudio
openai
wave
python-dotenv
elevenlabslib